+++ case_y_channel.f08
@@ -35,19 +35,19 @@ contains
 
         character(64) :: subroutine_name = 'case_y_channel_initialize'
 
-        real(4),  intent(in)  :: channel_length(:), channel_breadth(:)
-        real(4),  intent(in)  :: channel_topwidth(:), subdivide_length(:)
-        real(4),  intent(in)  :: lowerZ(:), upperZ(:),  initial_flowrate(:)
-        real(4),  intent(in)  :: depth_upstream(:), depth_dnstream(:), init_depth(:)
-        real(4),  intent(in)  :: ManningsN(:)
+        real(8),  intent(in)  :: channel_length(:), channel_breadth(:)
+        real(8),  intent(in)  :: channel_topwidth(:), subdivide_length(:)
+        real(8),  intent(in)  :: lowerZ(:), upperZ(:),  initial_flowrate(:)
+        real(8),  intent(in)  :: depth_upstream(:), depth_dnstream(:), init_depth(:)
+        real(8),  intent(in)  :: ManningsN(:)
 
         integer, intent(in):: roughness_type, idepth_type(:)
 
         integer,   dimension(:,:), allocatable, target, intent(out)    :: linkI
         integer,   dimension(:,:), allocatable, target, intent(out)    :: nodeI
 
-        real(4),      dimension(:,:), allocatable, target, intent(out)    :: linkR
-        real(4),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: linkR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
 
         logical,   dimension(:,:), allocatable, target, intent(out)    :: linkYN
         logical,   dimension(:,:), allocatable, target, intent(out)    :: nodeYN
@@ -117,18 +117,18 @@ contains
         !
         character(64) :: subroutine_name = 'case_simple_channel_links_and_nodes'
 
-        real(4),  intent(in)  :: channel_length(:), channel_breadth(:)
-        real(4),  intent(in)  :: channel_topwidth(:), subdivide_length(:)
-        real(4),  intent(in)  :: lowerZ(:), upperZ(:), ManningsN(:), initial_flowrate(:)
-        real(4),  intent(in)  :: depth_upstream(:), depth_dnstream(:), init_depth(:)
+        real(8),  intent(in)  :: channel_length(:), channel_breadth(:)
+        real(8),  intent(in)  :: channel_topwidth(:), subdivide_length(:)
+        real(8),  intent(in)  :: lowerZ(:), upperZ(:), ManningsN(:), initial_flowrate(:)
+        real(8),  intent(in)  :: depth_upstream(:), depth_dnstream(:), init_depth(:)
 
         integer, intent(in):: roughness_type, idepth_type(:)
 
         integer,   dimension(:,:), allocatable, target, intent(out)    :: linkI
         integer,   dimension(:,:), allocatable, target, intent(out)    :: nodeI
 
-        real(4),      dimension(:,:), allocatable, target, intent(out)    :: linkR
-        real(4),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: linkR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
 
         logical,   dimension(:,:), allocatable, target, intent(out)    :: linkYN
         logical,   dimension(:,:), allocatable, target, intent(out)    :: nodeYN
